Put 'move_by' back into Object.
This departs from the tutorial, but that felt pretty icky -- move_by
is a thing an object should handle for itself.

The block-check doesn't necessarily have to be a method on an object,
as it could be of interest to a lot of things whether a specific tile
is blocked, and it is an interaction between objects and the map, so I
opted to create a method that returns (dx, dy) if map[x+dx][y+dy] is
not blocked, and returns (0, 0) otherwise.